Change client-side limit of max image size
Changed the 600x400 image size to 3000x2000 because text and other small details are illegible on that resolution. 3000x2000 is just an example resolution that keeps the 3:2 aspect ratio, though I believe that 16:9 at 2560x1440 is better due to the prevalence of 16:9 images like screenshots. Preview in chat is too big though, but I think that it is a worth-it tradeoff for legible images.